Add new desktop places icon
The desktop icon is one of the few places icons
that has not seen a refresh, leaving the current
desktop icon looking a little bit dated.

This one is  brighter, cleaner, and better matches
the folder colors, gradients, outlines, and curves.
The sizing has also been adjusted to match the width
and height of the folders. This should also blend in
better with the newly designed Xfce app icons.