Remove and replace start-here
In the main theme start-here is the old elementary "e" logo.
In the dark theme start-here is the round Xubuntu logo.

This should be consistent between variants and probably shouldn't
represent a specific distribution or desktop environment since
it can be used as a generic menu icon.

This PR removes the existing icons and replaces them with a generic
grid-style icon.