Deprecate elementary Xfce darker
With the updates over the last year to reduce the need for the
fake monochromatic toolbar icons, they should be safe to remove,
removing the need for another variant.

The light and dark panel icons do still seem needed in some places,
so they remain. This will leave the theme with two variants:
"elementary Xfce" and "elementary Xfce dark".

Similar to what was done with the 'darkest' variant, a
deprecation message was added to the description of darker
telling users to now use 'elementary Xfce dark'

Fixes #346

These may have been fixed already, but if not will be fixed with this:

Fixes #203
Fixes #216