Remove notification- prefixed icons
Remove most of the notification- prefixed icons. The only mention
of these names seems to be in Ubuntu's notify-osd, which, while
still in the repository has not been used in Ubuntu since 14.04,
and isn't used in any of the Ubuntu flavors. notify-osd is not
in the Fedora, OpenSuse, etc reposities.

Remaining icons in the notification folder were moved to the status folder.
notification folder was removed, entries removed from index.theme.

nm-applet seemed to use a few of the Network notification- icons, so those were left.

Updated krb- icons to use newer elementary status symbols.

Added gnome-lockscreen symlink which is used for nm-applet notification
when connecting to or disconnecting from a VPN.