Apps: Remove (most likely) unused symlinks
Some of these are definitely no longer used, but a few
of the very generically named icons (fonts) are difficult
to search for. Did my best to look for them and compare
them to other major icon themes. They don't seem to be
provided by any major themes and I think are safe for removal.

Removed:
- addressbook
- background
- browser-help
- fonts
- hotwire (hotwire-shell, no development since 2008)
- office-address-book (renamed to freedesktop.org name mimes/x-office-address-book)
- pirut (not in RHEL/Fedora since 2008)
- pk-action-refresh-cache (removed from packagekit in 2008)
- pk-package-sources (removed from packagekit in 2008)
- porthole (old gentoo portage frontend no dev since 2010, removed from repos in 2020)
- preferences-wallpaper (preferences-desktop-wallpaper is used instead)
- screensaver
- stock_* symlinks (removed like in #304, etc. Left a few still being used by evolution)
- susehelpcenter (removed since Suse 15)
- wallpaper