Mimes: Use 'applications-other' for executables
Symlink the Mime icon `application-x-executable` to the
Category icon `applications-other`. It uses the same design
as what is used currently but has better contrast, looks better
on dark themes, and is used more widely in the theme to represent applications.

This will change the icon used for executable and binary mimes, as well as some
executable-related actions like "Create Launcher..." on Xfdesktop right click.

---

The `applications-other` icon is used for:
- "All Applications" in Whisker Menu
- "Launcher" Xfce panel plugin
- "Applications" when right clicking on Xfdesktop

The white-square-with-gear is also used for .desktop files on Xfdesktop
and thunar, and also more closely matches the icon used for other executable
mimes like Appimage, Debs, RPMs, etc.