Add HiDPI scaling support for high resolution displays
Add an optional --hidpi configure flag to allow installing
as a fully SVG theme, with an updated index.theme for HiDPI
support. This will be seen when Scaling is set to 2x or 3x.

Fixes issues where wrong icons are shown in some places with
scaling enabled (like the Places sidebars in file managers),
and blurry icon issues, especially with scaling enabled.
The PNG non-HiDPI theme also restricted the amount of zoom
steps the icons had with scaling enabled, so this should
resolve that as well.

Fixes #124
Fixes #317
Fixes #362
Fixes #403