Add missing action, status icons for FreeDesktop
Add missing action and status icons to better comply
with the FreeDesktop Icon Naming Specification.

Elements of the weather icons are symbolic recolors,
may need to add borders and adjust colors in the future.

https://specifications.freedesktop.org/icon-naming-spec/icon-naming-spec-latest.html

Like the last "add icons for FD.o" commit, I don't know of places where
these are used, but better to be closer to the spec than not.