Add HiDPI theme for better scaling, hi-res displays
Install SVG-based HiDPI theme alongside the PNG-based theme.
This should be used for better icon scaling, specifically when
Scaling is set to 2x or 3x. PNG may be a bit faster to load,
so the SVG HiDPI theme is optional.

Using the HiDPI theme will fix issues where the wrong icons are shown in
some places with scaling enabled (like the Places sidebars in file managers).
It should also fix and blurry icon issues, especially with scaling enabled.
The PNG non-HiDPI theme restricted the amount of zoom steps the icons had
in file managers with scaling enabled, so this should resolve that as well.

Fixes #124
Fixes #317
Fixes #362
Fixes #403