Apps: Use new folder style for Catfish
This is the icon upstream uses for `system-file-manager`.

But elementary-xfce has always used folder-with-home-emblem
for `system-file-manager` and folder-with-magnifying-glass-emblem
for Catfish. So this just continues that style.

Maybe in the future following upstream will make more sense,
in which case maybe a new design for Catfish will be needed
(or not providing a Catfish icon and defaulting to the app-provided icon).