Fix workflow not building icon theme correctly
There seems to be some issue with the png to svg
conversion in the github actions workflow when
using `ubuntu-latest` (which is currently Ubuntu 24.04).
This is resolved by using `ubuntu-22.04`.

Not sure why this happens, I've tested on Ubuntu 24.04 and
do not get the build errors.

This fix feels like a band-aid but it works for now.