Dark: Add deprecation warning
With the move to colorize and unify these icon themes,
the -dark variant is no longer needed or used.

This adds a deprecation warning but leaves the folder
and index.theme installed so that a user's icon theme
will not abruptly change. It simply falls back to
elementary-xfce. At some point in the future, after
people have had time to switch to the main theme in
their settings, the -dark theme will be removed.