MimeTypes: Add text to 32px archive icons
Tries to use the text from the 48px icons untouched.
Proportionally the text is larger compared to the icon,
but keeping the 48px text size makes it more legible.

A few had to be made a bit smaller to fit, some
shortened (tar.xyz shortened to t.xyz for example).