MimeTypes: New design for broken symlink icon
The `inode-symlink` icon, in all major file managers,
only shows if a symlink is broken. Previously a trash
emblem was used since they're empty and safe to delete,
but it doesn't really symbolize "broken symlink" and it's
not always best to delete them or symbolize them as trash.

This new design uses the symlink emblem with the error
symbol to let the user know something is wrong. Hopefully
this is a more clear symbol.