Actions: Remove 96px, 128px action icons
These are very incomplete, and most of the ones
that exist are non up to date with the other sizes.

Action icons are generally used at 16-24px.
Rarely at 32px and even moreso at 48px.
Upstream removed 64px and above long ago.
But we'll keep 64px for now since there may be
cases where some DEs haven't implemented proper
HiDPI scaling, and may need the 48px and 64px
icons when 24px and 32px icons are doubled.

Removed some symlinks that turned out to be
not used or needed along the way:

- `messagebox_*` - Don't seem to be used anywhere
                   anymore, possibly old KDE3 icons
- `setroubleshoot_icon` - Fedora's setroubleshoot app has its own
                          newer/nicer icons (including scalable and
                          symbolic), ours doesn't match design anyway
- `stock_view-details` - Old Gtk2 icon, only can find use in Evolution
                         at 16px and they provide a better icon